requirements.txt
-i https://pypi.org/simple
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
contourpy==1.2.0; python_version >= '3.9'
cycler==0.12.1; python_version >= '3.8'
fonttools==4.48.1; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.6; python_version >= '3.5'
kiwisolver==1.4.5; python_version >= '3.7'
matplotlib==3.8.3; python_version >= '3.9'
numpy==1.26.4; python_version >= '3.12'
packaging==23.2; python_version >= '3.7'
pandas==2.2.0; python_version >= '3.9'
pillow==10.2.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyarrow==15.0.0; python_version >= '3.8'
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.27; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
urllib3==2.2.0; python_version >= '3.8'
